[client-web] Fixed rounding in resource capacity utilization.

+++ shongo-client-web/src/main/java/cz/cesnet/shongo/client/web/admin/ResourceCapacity.java
@@ -101,7 +101,7 @@ public abstract class ResourceCapacity
                     output.append(value);
                     break;
                 case RELATIVE:
-                    output.append((utilizedLicenseCount.intValue() * 100) / licenseCount);
+                    output.append((int)((utilizedLicenseCount.doubleValue() * 100.0) / (double) licenseCount));
                     output.append("%");
                     break;
                 default:
